#44: Code reorganization
