+++ src/module/consts/consts.js
@@ -9,9 +9,12 @@ export const SYSTEM_NAME = 'OQ System';
 const SettingsConfig = {
   keys: {
     defaultItemsCompendium: 'defaultSkillsCompendium',
+    coinsConfiguration: 'coins',
   },
   defaults: {
     characterItemsCompendium: 'oq.oq-system-basic-skills',
+    defaultCoinsConfiguration:
+      'Gold Ducat (GD) = 20, Silver Piece (SP) = 1, Copper Penny (CP) = 0.1, Lead Bit (LB) = 0.02',
   },
 };
