+++ src/module/document/item/base-item.js
@@ -1,6 +1,6 @@
-import { displayItem } from '../../chat.js';
+import { displayItem } from '../../utils/chat.js';
 import _ from 'lodash-es';
-import { damageRoll, testRoll } from '../../roll.js';
+import { damageRoll, testRoll } from '../../utils/roll.js';
 import { OQTestRollDialog } from '../../application/test-roll-dialog.js';
 import { OQDamageRollDialog } from '../../application/damage-roll-dialog.js';
