+++ src/module/init/custom-hook-handlers.js
@@ -1,6 +1,6 @@
 import { handleDamageRollChatMessage } from '../chat-handlers/updates-from-chat.js';
 import { commandHandler } from '../chat-handlers/chat-command-listener.js';
-import { createItemMacro } from '../item-macro.js';
+import { createItemMacro } from '../utils/item-macro.js';
 
 export function registerCustomHookHandlers() {
   Hooks.on('renderChatMessage', handleDamageRollChatMessage);
