+++ src/module/oq.js
@@ -1,12 +1,12 @@
 import { registerSettings } from './init/register-settings.js';
 import { preloadTemplates } from './init/preload-templates.js';
-import { log } from './utils.js';
 import { OQ } from './consts/consts.js';
 import { registerDocuments } from './init/register-documents.js';
 import { registerDataModels } from './init/register-data-models.js';
 import { registerHelpers } from './init/handlebar-helpers.js';
 import { registerCustomHookHandlers } from './init/custom-hook-handlers.js';
-import { oqGame } from './oq-game.js';
+import { oqGame } from './utils/oq-game.js';
+import { log } from './utils/logger.js';
 
 async function init() {
   log('Initializing OQ');
