+++ src/module/utils/compendium-utils.js
@@ -23,6 +23,5 @@ export async function getDefaultItemsForActor(actorType) {
       }
     }
   }
-
   return [];
 }
