+++ src/module/utils/oq-game.js
