+++ src/module/utils/utils.js
@@ -1,19 +1,4 @@
 import _ from 'lodash-es';
-const SYSTEM_NAME = 'OQ System';
-export const log = console.log.bind(undefined, `${SYSTEM_NAME} |`);
-
-export const logError = console.error.bind(undefined, `${SYSTEM_NAME} |`);
-
-/**
- *
- * @param {string} message
- * @param {Object|string|number|boolean|undefined} object
- * @returns {*}
- */
-export const logObject = (message, object) => {
-  log(message, object);
-  return object;
-};
 
 /**
  * Number as string with sign. In case of 0 it returns empty string.
