+++ test/test-roll.js
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import { getResult, getResultFeatures } from '../src/module/roll.js';
+import { getResult, getResultFeatures } from '../src/module/utils/roll.js';
 import { RollConfig } from '../src/module/consts/rolls-config.js';
 
 describe('roll.js', function () {
